consumer_product_filter_api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.8.5
Pillow==10.4.0
requests==2.32.4
